requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
Django==5.2.5
django-js-asset==3.1.2
django-mptt==0.17.0
django-ratelimit==4.1.0
django-widget-tweaks==1.5.0
ghasedak_sms==1.0.3
ghasedakpack==0.1.13
idna==3.10
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0

django-environ~=0.12.0
django-colorfield~=0.14.0